test: Renamed test files
